change config: now info to error level logs are stored in info.log file and debug to error level logs are displayed at console
